--- .pylintrc
@@ -1,15 +1,0 @@
-[master]
-load-plugins=pylint.extensions.mccabe, pylint.extensions.docparams
-
-[FORMAT]
-max-line-length=100
-
-[MESSAGES CONTROL]
-disable=bad-continuation
-
-[DESIGN]
-max-parents=13
-max-complexity=10
-
-[TYPECHECK]
-generated-members=REQUEST,acl_users,aq_parent,"[a-zA-Z]+_set{1,2}",save,delete
