--- config.local.py
@@ -1,5 +1,0 @@
-"""Configuration file for Local Development"""
-
-# If true page modules will be reloaded before their write function is applied
-# See
-DEBUG = False
