--- config.prod.py
@@ -1,2 +1,0 @@
-"""Configuration file for production Development"""
-DEBUG = False
