--- config.py
@@ -1,2 +1,0 @@
-"""Configuration file for Local Development"""
-DEBUG = True
