+++ home.py
@@ -1,7 +1,6 @@
 """Main module for the streamlit app"""
 import streamlit as st
 from PIL import Image
-#from st_pages import Page, Section, show_pages, add_page_title
 
 primaryColor="#264653"
 backgroundColor="#FFFFFF"
@@ -15,25 +14,6 @@ st.set_page_config(
       layout="wide",
       initial_sidebar_state="collapsed",
 )
-
-#import awesome_streamlit as ast
-
-#show_pages(
-#    [
-#        Page("home.py", "Home", "🏠"),
-#        Page("pages/plan.py", "Plan", ":spiral_note_pad:"),
-#        Page("pages/saveplan.py", "Save", ":left_speech_bubble:"),
-#        Section(name="Reports", icon=":books:"),
-#        Page("pages/stoplight.py", "StopLight", ":books:"),
-#        Page("pages/canvas.py", "Canvas", ":books:"),
-#        Page("pages/finalreport.py", "  Final", ":information_source:"),
-#        Section(name="Analysis", icon=":books:"),
-#        Page("pages/risk.py", "Risk", ":books:"),
-#        Page("pages/wbs.py", "WBS", ":books:"),
-#        Page("pages/communication.py", "  Comms", ":left_speech_bubble:"),
-#        Page("pages/about.py", "About", ":books:", in_section=False),
-#    ]
-#)
 
 def main():
     """Main function of the App"""
@@ -86,23 +66,6 @@ Closure \n
     with colr[1]:
      st.image(ip, caption='Step 3 - The PM Monitor Closing')
     st.markdown("---")
-    columns = st.columns((2,1,2))
-    url1 = "/plan"
-#    with columns[1]:
-#     st.markdown(f'''
-#<a href={url1} target = "_self"><button style="background-color:#F4A261;text-align: center;">  Start Planning  </button></a>
-#''',
-#unsafe_allow_html=True)
-
-#    st.sidebar.write("***About and Contribute***")
-#    st.sidebar.write(
-#        """
-#        This an open source project built using Streamlit, submit comments, questions, as 
-#        [issues](https://github.com/jlastwood/pmstreamlit/issues)
-#        Idea, Design and Development by Janet. More information at
-#        [bluezoneit.com](https://bluezoneit.com).  Feel free to buy me a coffee to support The PM monitor.  [bmc.link](https://bmc.link/thepmmonitor)
-#"""
-#    )
 
 if __name__ == "__main__":
     main()
