--- pages/requirements.txt
@@ -1,255 +1,0 @@
-aiohttp==3.7.2
-alembic==1.5.4
-altair==4.1.0
-anyascii==0.2.0
-appdirs==1.4.4
-appnope==0.1.2
-argon2-cffi==21.1.0
-arrow==0.17.0
-astor==0.8.1
-async-timeout==3.0.1
-attrs==20.3.0
-backcall==0.2.0
-backports.zoneinfo==0.2.1
-base58==2.1.0
-basecampy3==0.7.2
-bcrypt==3.2.0
-Beaker==1.11.0
-beautifulsoup4==4.9.3
-binaryornot==0.4.4
-black==20.8b1
-bleach==4.1.0
-blinker==1.4
-cachetools==4.2.4
-certifi==2020.11.8
-cffi==1.14.5
-Chameleon==3.9.1
-chardet==3.0.4
-click==7.1.2
-colander==1.8.3
-cookiecutter==1.7.2
-debugpy==1.5.0
-decorator==5.1.0
-deform==2.0.14
-defusedxml==0.7.1
-distlib==0.3.1
-dnspython==2.1.0
-email-validator==1.1.2
-entrypoints==0.3
-environs==9.3.1
-factory-boy==3.2.0
-Faker==6.1.1
-fancycompleter==0.9.1
-fanstatic==1.1
-filelock==3.0.12
-flake8==3.8.4
-flake8-blind-except==0.2.0
-flake8-debugger==4.0.0
-flake8-docstrings==1.5.0
-flake8-isort==4.0.0
-flake8-polyfill==1.0.2
-Flask==1.1.2
-Flask-Bcrypt==0.7.1
-Flask-Caching==1.9.0
-Flask-DebugToolbar==0.11.0
-Flask-FontAwesome==0.1.4
-Flask-Login==0.5.0
-Flask-Markdown==0.3
-Flask-Migrate==2.6.0
-Flask-PageDown==0.3.0
-Flask-SQLAlchemy==2.4.4
-Flask-Static-Digest==0.2.1
-Flask-WTF==0.14.3
-fontawesome==5.10.1.post1
-forex-python==1.5
-FormEncode==2.0.0
-future==0.18.2
-gevent==21.1.2
-gitdb==4.0.7
-GitPython==3.1.24
-greenlet==1.0.0
-gunicorn==20.0.4
-html2text==2020.1.16
-hupper==1.10.3
-idna==2.10
-importlib-metadata==4.11.3
-iniconfig==1.1.1
-ipykernel==6.4.1
-ipython==7.28.0
-ipython-genutils==0.2.0
-ipywidgets==7.6.5
-iso8601==0.1.14
-isort==5.7.0
-itsdangerous==1.1.0
-jedi==0.18.0
-Jinja2==2.11.3
-jinja2-time==0.2.0
-joblib==0.16.0
-js.angular==1.1.4
-js.bootstrap==3.3.4
-js.deform==2.0.14
-js.fineuploader==5.14.0
-js.html5shiv==3.7.3
-js.jquery==1.9.1
-js.jquery-form==3.9
-js.jquery-maskedinput==1.4.1
-js.jquery-maskmoney==3.0.2
-js.jquery-sortable==0.9.13
-js.jquery-tablednd==0.4
-js.jquery-timepicker-addon==1.5.3
-js.jqueryui==1.10.3
-js.jqueryui-tagit==2.0.24.post2
-js.modernizr==2.5.3.1
-js.select2==4.0.4.post1
-js.tinymce==4.5.4
-jsonschema==4.0.1
-jupyter-client==7.0.6
-jupyter-core==4.8.1
-jupyterlab-pygments==0.1.2
-jupyterlab-widgets==1.0.2
-lingua==4.14
-Mako==1.1.4
-Markdown==3.3
-markdown-it-py==2.2.0
-MarkupSafe==1.1.1
-marshmallow==3.10.0
-matplotlib-inline==0.1.3
-mccabe==0.6.1
-mdurl==0.1.2
-mistune==0.8.4
-multidict==5.0.0
-mypy-extensions==0.4.3
-nbclient==0.5.4
-nbconvert==6.2.0
-nbformat==5.1.3
-nest-asyncio==1.5.1
-nltk==3.5
-notebook==6.4.4
-numpy==1.24.3
-oauthlib==3.1.0
-packaging==20.9
-pandas==1.0.5
-pandocfilters==1.5.0
-parso==0.8.2
-PasteDeploy==2.1.1
-pathspec==0.8.1
-pdbpp==0.10.2
-pep8-naming==0.11.1
-peppercorn==0.6
-pexpect==4.8.0
-pick==0.6.7
-pickleshare==0.7.5
-Pillow==8.3.2
-pipenv==2020.11.4
-plaster==1.0
-plaster-pastedeploy==0.7
-pluggy==0.13.1
-polib==1.1.1
-poyo==0.5.0
-prometheus-client==0.11.0
-prompt-toolkit==3.0.20
-protobuf==4.23.2
-psycopg2==2.8.3
-psycopg2-binary==2.8.6
-ptyprocess==0.7.0
-py==1.10.0
-py-bcrypt==0.4
-py-trello==0.16.0
-pyarrow==5.0.0
-pycodestyle==2.6.0
-pycparser==2.20
-pydeck==0.7.0
-pydocstyle==5.1.1
-pyflakes==2.2.0
-Pygments==2.15.1
-Pympler==1.0.1
-pyparsing==2.4.7
-pyramid==1.10.8
-pyramid-beaker==0.8
-pyramid-chameleon==0.3
-pyramid-deform==0.2
-pyramid-mailer==0.15.1
-pyramid-tm==2.4
-pyramid-zcml==1.2.0
-pyrepl==0.9.0
-pyrsistent==0.18.0
-pystache==0.5.4
-pytest==6.2.2
-python-dateutil==2.8.1
-python-dotenv==0.15.0
-python-editor==1.0.4
-python-slugify==4.0.1
-pytz==2020.1
-pyzmq==22.3.0
-redis==3.5.3
-redis-server==6.0.9
-regex==2020.11.13
-repoze.lru==0.7
-repoze.sendmail==4.4.1
-repoze.workflow==1.1
-repoze.zcml==1.1
-requests==2.24.0
-requests-oauthlib==1.3.0
-rich==13.4.1
-rq==1.5.2
-semver==2.13.0
-Send2Trash==1.8.0
-shutilwhich==1.1.0
-simplejson==3.17.2
-six==1.15.0
-slack==0.0.2
-slackclient==2.9.3
-slacker==0.14.0
-smmap==4.0.0
-snowballstemmer==2.1.0
-soupsieve==2.2
-SQLAlchemy==1.3.23
-SQLAlchemy-Utils==0.37.6
-st-pages==0.4.1
-streamlit==1.23.1
-streamlit-metrics==0.1.0
-supervisor==4.2.1
-tenacity==8.2.2
-terminado==0.12.1
-testfixtures==6.17.1
-testpath==0.5.0
-text-unidecode==1.3
-toml==0.10.2
-toolz==0.11.1
-tornado==6.1
-tqdm==4.48.0
-traitlets==5.1.0
-transaction==3.0.1
-translationstring==1.4
-trello==0.9.4
-typed-ast==1.4.2
-typing-extensions==4.6.3
-tzlocal==2.1
-Unidecode==1.2.0
-urllib3==1.25.9
-vaderSentiment==3.3.2
-validators==0.18.2
-venusian==3.0.0
-virtualenv==20.1.0
-virtualenv-clone==0.5.4
-waitress==1.4.4
-wcwidth==0.2.5
-webencodings==0.5.1
-WebOb==1.8.6
-WebTest==2.0.35
-Werkzeug==1.0.1
-widgetsnbextension==3.5.1
-wmctrl==0.3
-WTForms==2.3.3
-XlsxWriter==1.2.9
-yarl==1.6.2
-zipp==3.8.0
-zope.component==5.0.0
-zope.configuration==4.4.0
-zope.deprecation==4.4.0
-zope.event==4.5.0
-zope.hookable==5.0.1
-zope.i18nmessageid==5.0.1
-zope.interface==5.4.0
-zope.schema==6.1.0
-zope.sqlalchemy==1.4
