+++ scripts/askchatgpt.py
@@ -1,5 +1,4 @@
 import streamlit as st
-import re
 import json
 import requests
 #import openai
