+++ scripts/experiments/madsen_stuff/.dockerignore
