+++ scripts/experiments/take_screenshots_of_resources.py
