--- scripts/extract_resources_markdown.py
@@ -1,3 +1,0 @@
-import awesome_streamlit.core.services.resources as resources
-
-print(resources.get_resources_markdown([]))
