--- scripts/hello_streamlit.py
@@ -1,3 +1,0 @@
-import streamlit as st
-
-st.title("Hello Streamlit again and again")
